requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types
antlr4-python3-runtime==4.9.3
anyio
appdirs==1.4.4
asgiref
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
basicsr==1.4.2
Brotli
cached-property==1.5.2
cachetools==5.3.1
certifi
cffi
chardet==5.2.0
charset-normalizer
chex
click==8.1.7
cmake==3.28.1
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.3.0
django-extensions==3.2.3
djangorestframework==3.14.0
dlib==19.24.2
dm-tree==0.1.8
docker-pycreds==0.4.0
easydict==1.11
einops==0.7.0
etils==1.3.0
exceptiongroup==1.2.0
executing==2.0.0
facexlib==0.3.0
fastprogress==1.0.3
ffmpeg-python==0.2.0
filelock==3.12.4
filterpy==1.4.5
flatbuffers==23.5.26
flax
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
# Editable install with no version control (gfpgan==1.3.8)
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
huggingface-hub
hydra-core==1.3.2
idna
imageio==2.33.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipdb==0.13.13
ipython
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.1
joblib
keras==2.13.1
kiwisolver==1.4.5
kornia==0.6.7
lazy_loader==0.3
libclang==16.0.6
lightning
lightning-utilities
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft
mkl-random
mkl-service==2.4.0
ml-collections==0.1.1
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
mutagen==1.47.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numpy
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint
packaging
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs
pooch
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser
pycryptodomex==3.19.0
pydantic==1.10.14
pydantic_core==0.42.0
pyDeprecate==0.3.2
Pygments==2.16.1
pyOpenSSL
pyparsing==3.1.1
PySocks
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning
pytorch-msssim==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
realesrgan==0.3.0
requests
requests-oauthlib==1.3.1
rich
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn
scipy==1.9.1
sentry-sdk==1.32.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
starlette
sympy==1.12
tb-nightly==2.14.0a20230808
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorstore==0.1.45
termcolor==2.4.0
threadpoolctl
tifffile==2023.7.10
tiler==0.5.7
tiling==0.3.0
timm==0.9.2
tomli==2.0.1
toolz==0.12.1
torch==1.13.1
torchaudio==0.13.1
torchdata==0.5.1
torchmetrics==0.11.4
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.65.0
traitlets==5.11.2
typing==3.7.4.3
typing_extensions
tzdata==2023.3
urllib3
wandb==0.15.5
wcwidth==0.2.8
websockets==11.0
Werkzeug==3.0.0
wrapt==1.16.0
yapf==0.40.2
yarl==1.9.2
yt-dlp==2023.10.13
zipp==3.17.0
